update dependencies and rename distributed cache
