+++ src/api/pyproject.toml
@@ -12,6 +12,13 @@ langserve = "^0.2.2"
 uvicorn = "^0.23.2"
 pydantic_settings = "^2.3.4"
 redis = "^5.0.7"
+models = { path = "../models", develop = true }
+utils = { path = "../utils", develop = true }
+agents = { path = "../agents", develop = true }
+distributedcache = { path = "../distributed_cache", develop = true }
+services = { path = "../services", develop = true }
+fastapi = "^0.112.0"
+langchain-community = "^0.2.11"
 
 [tool.poetry.group.dev.dependencies]
 env = "^0.1.0"
