+++ src/services/pyproject.toml
@@ -8,7 +8,7 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = "^3.11"
-distributed_cache = { path = "../distributed_cache", develop = true }
+distributedcache = { path = "../distributed_cache", develop = true }
 
 [build-system]
 requires = ["poetry-core"]
